partial completion of store and item
