+++ app/src/main/AndroidManifest.xml
@@ -26,8 +26,7 @@
         </activity>
         <activity android:name=".controllers.ApplicationActivity" />
         <activity android:name=".controllers.ListViewCustom" />
-        <activity
-            android:name=".controllers.ListElementPage"></activity>
+        <activity android:name=".controllers.ListElementPage" />
         <activity
             android:name=".controllers.RegistrationActivity"
             android:label="@string/title_activity_registration"
@@ -38,7 +37,10 @@
         </activity>
         <activity
             android:name=".controllers.StoreViewActivity"
-            android:label="@string/title_activity_store_view"></activity>
+            android:label="@string/title_activity_store_view"/>
+        <activity
+            android:name=".controllers.ItemView"
+            android:label="@string/title_activity_item_view"/>
     </application>
 
 </manifest>
