+++ app/src/main/java/com/scrumbums/donationboi/model/util/AccountValidation.java
@@ -18,20 +18,9 @@ public final class AccountValidation {
     private static final String EMAILREGEX = "(\\w|\\d|\\.|_)+@(\\w|\\d|\\.|_)+\\.(\\w|\\d)+";
 
     /**
-     * Special characters list. Used in constructing the following
-     * regular expressions.
-     */
-    private static final String SPECIALCHARS = "@_\\$~\\.!-\\^%";
-
-    /**
-     * Regular expression for 'special characters.'
-     */
-    private static final String SPECIALCHARSREGEX = "[" + SPECIALCHARS + "]+";
-
-    /**
      * Regular expression for legal characters.
      */
-    private static final String LEGALCHARSREGEX = "[a-zA-Z0-9" + SPECIALCHARS + "]+";
+    private static final String LEGALCHARSREGEX = "[a-zA-Z0-9]+";
 
     /**
      * Minimum required password length.
@@ -47,11 +36,6 @@ public final class AccountValidation {
      * Minimum number of lowercase characters needed in password.
      */
     private static final int MINLOWERCASECHARS = 3;
-
-    /**
-     * Minimum number of special characters required in a password.
-     */
-    private static final int MINSPECIALCHARS = 1;
 
     /**
      * Minimum number of numeric characters required in a password.
@@ -95,8 +79,9 @@ public final class AccountValidation {
      * @return If the String is a valid password
      */
     public static final boolean isValidPassword(String s) {
-        return isStringNonEmpty(s) && !containsIllegalCharacters(s)
-                && s.length() >= MINPASSWORDLENGTH && hasMinChars(s);
+        return true;
+//        return isStringNonEmpty(s) && !containsIllegalCharacters(s)
+//                && s.length() >= MINPASSWORDLENGTH && hasMinChars(s);
     }
 
     /**
@@ -108,13 +93,10 @@ public final class AccountValidation {
     private static final boolean hasMinChars(String s) {
         int numUpper = 0;
         int numLower = 0;
-        int numSpecial = 0;
         int numNum = 0;
         for (int i = 0; i < s.length(); i++) {
             char c = s.charAt(i);
-            if ((c + "").matches(SPECIALCHARSREGEX)) {
-                numSpecial++;
-            } else if (c >= 'a' && c <= 'z') {
+            if (c >= 'a' && c <= 'z') {
                 numLower++;
             } else if (c >= 'A' && c <= 'Z') {
                 numUpper++;
@@ -123,6 +105,6 @@ public final class AccountValidation {
             } // else should never happen
         }
         return numUpper >= MINUPPERCASECHARS && numLower >= MINLOWERCASECHARS
-                && numSpecial >= MINSPECIALCHARS && numNum >= MINNUMERICCHARS;
+                && numNum >= MINNUMERICCHARS;
     }
 }
